Fix issues from static analysis (#202)
* Casting a signed value which has type 'int' to a bigger unsigned integer type 'size_t' while initializing a variable
* Assignment of a signed value which has type 'int32_t' to a variable of a bigger integer type 'size_t'
* Uninitialized field

+++ shell/platform/tizen/channels/text_input_channel.cc
@@ -267,8 +267,8 @@ void TextInputChannel::HandleMethodCall(
     if (composing_base_value == -1 && composing_extent_value == -1) {
       active_model_->EndComposing();
     } else {
-      size_t composing_start =
-          std::min(composing_base_value, composing_extent_value);
+      size_t composing_start = static_cast<size_t>(
+          std::min(composing_base_value, composing_extent_value));
       size_t cursor_offset = selection_base_value - composing_start;
 
       active_model_->SetComposingRange(
